+++ src/utilities/typeCheckers.js
@@ -1,0 +1,5 @@
+function isFloat(value) {
+    return !isNaN(parseFloat(value)) && isFinite(value);
+}
+
+module.exports = { isFloat };